+++ dashboard/views.py
@@ -11,7 +11,6 @@ import pandas as pd
 from dashboard.models import (
     QuestionArchive,
     Question,
-    User,
     Answer,
     UncuratedSubmission,
     UnencodedSubmission,
@@ -19,30 +18,20 @@ from dashboard.models import (
 from pprint import pprint
 
 
-def get_login_view(request):
-    """Return the login view."""
-    return render(request, 'dashboard/login.html')
-
-
-def get_signup_view(request):
-    """Return the signup view."""
-    return render(request, 'dashboard/signup.html')
-
-
-@login_required(login_url='login/')
+@login_required
 def get_home_view(request):
     """Return the dashboard home view."""
     context = {}
     return render(request, 'dashboard/home.html', context)
 
 
-@login_required(login_url='login/')
+@login_required
 def get_submit_questions_view(request):
     """Return the view for submitting questions."""
     return render(request, 'dashboard/submit-questions.html')
 
 
-@login_required(login_url='login/')
+@login_required
 def get_submit_excel_sheet_view(request):
     """Return the view for submitting Excel sheet."""
     context = {
@@ -51,7 +40,7 @@ def get_submit_excel_sheet_view(request):
     return render(request, 'dashboard/submit-excel-sheet.html', context)
 
 
-@login_required(login_url='login/')
+@login_required
 def get_view_questions_view(request):
     """Return the 'View Questions' view after applying filters, if any."""
     questions_superset = Question.objects.all().order_by('-created_on')
@@ -75,7 +64,7 @@ def get_view_questions_view(request):
     return render(request, 'dashboard/view-questions.html', context)
 
 
-@login_required(login_url='login/')
+@login_required
 def get_answer_questions_list_view(request):
     """Return the view with list of unanswered questions"""
 
@@ -88,7 +77,7 @@ def get_answer_questions_list_view(request):
     return render(request, 'dashboard/answer-questions-list.html', context)
 
 
-@login_required(login_url='login/')
+@login_required
 def get_answer_question_view(request, question_id):
     """Return the view to answer a question"""
 
@@ -100,7 +89,7 @@ def get_answer_question_view(request, question_id):
     return render(request, 'dashboard/answer-question.html', context)
 
 
-@login_required(login_url='login/')
+@login_required
 def get_curate_data_view(request):
     """Return the curate data view"""
 
@@ -115,7 +104,7 @@ def get_curate_data_view(request):
     return render(request, 'dashboard/curate-data.html', context)
 
 
-@login_required(login_url='login/')
+@login_required
 def get_encode_data_view(request):
     """Return the encode data view"""
 
@@ -130,68 +119,68 @@ def get_encode_data_view(request):
     return render(request, 'dashboard/encode-data.html', context)
 
 
-def login_user(request):
-    """Log the user in"""
+# def login_user(request):
+#     """Log the user in"""
 
-    email = request.POST.get('email')
-    password = request.POST.get('password')
-    user = authenticate(request, email=email, password=password)
+#     email = request.POST.get('email')
+#     password = request.POST.get('password')
+#     user = authenticate(request, email=email, password=password)
 
-    if user is not None:
-        login(request, user)
-        return redirect(request.POST.get('next', 'dashboard:dashboard_home'))
-    else:
-        return render(
-            request,
-            'dashboard/login.html',
-            {'error': 'Incorrect login info! Please try again'})
-
-
-def logout_user(request):
-    """Log out the user"""
-
-    logout(request)
-    return redirect('dashboard:dashboard_home')
+#     if user is not None:
+#         login(request, user)
+#         return redirect(request.POST.get('next', 'dashboard:dashboard_home'))
+#     else:
+#         return render(
+#             request,
+#             'dashboard/login.html',
+#             {'error': 'Incorrect login info! Please try again'})
 
 
-def signup_user(request):
-    """Create a user"""
+# def logout_user(request):
+#     """Log out the user"""
 
-    if request.POST.get('password') != request.POST.get('re-password'):
-        return render(
-            request,
-            'dashboard/signup.html',
-            {'error': 'Passwords do not match! Please try again'})
+#     logout(request)
+#     return redirect('dashboard:dashboard_home')
 
-    email = request.POST.get('email').strip()
-    email_exists = User.objects.filter(email=email).exists()
 
-    if email_exists:
-        return render(
-            request,
-            'dashboard/signup.html',
-            {'error': 'Email already exists! Try logging in!'})
-    else:
-        password = request.POST.get('password').strip()
-        organisation = request.POST.get('organisation').strip()
-        access_requested = ','.join(request.POST.getlist('access_request'))
+# def signup_user(request):
+#     """Create a user"""
 
-        user = User.objects.create_user(email, password)
-        user.first_name = request.POST.get('first_name').strip()
-        user.last_name = request.POST.get('last_name').strip()
+#     if request.POST.get('password') != request.POST.get('re-password'):
+#         return render(
+#             request,
+#             'dashboard/signup.html',
+#             {'error': 'Passwords do not match! Please try again'})
 
-        if organisation == 'other':
-            organisation = request.POST.get('other-org')
+#     email = request.POST.get('email').strip()
+#     email_exists = User.objects.filter(email=email).exists()
 
-        user.organisation = organisation
-        user.access_requested = access_requested
-        user.save()
+#     if email_exists:
+#         return render(
+#             request,
+#             'dashboard/signup.html',
+#             {'error': 'Email already exists! Try logging in!'})
+#     else:
+#         password = request.POST.get('password').strip()
+#         organisation = request.POST.get('organisation').strip()
+#         access_requested = ','.join(request.POST.getlist('access_request'))
 
-        volunteers = Group.objects.get(name='volunteers')
-        volunteers.user_set.add(user)
+#         user = User.objects.create_user(email, password)
+#         user.first_name = request.POST.get('first_name').strip()
+#         user.last_name = request.POST.get('last_name').strip()
 
-        login(request, user)
-        return redirect('dashboard:dashboard_home')
+#         if organisation == 'other':
+#             organisation = request.POST.get('other-org')
+
+#         user.organisation = organisation
+#         user.access_requested = access_requested
+#         user.save()
+
+#         volunteers = Group.objects.get(name='volunteers')
+#         volunteers.user_set.add(user)
+
+#         login(request, user)
+#         return redirect('dashboard:dashboard_home')
 
 
 def submit_questions(request):
