+++ sawaliram_auth/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class SawaliramAuthConfig(AppConfig):
+    name = 'sawaliram_auth'
